chore: 调整创建动态壁纸相关逻辑
