+++ electron/create-mac-live-wallpaper.ts
@@ -7,7 +7,7 @@ async function requireWallpaper() {
 
 let wallWindow: BrowserWindow[] = []
 // 创建窗口
-export function createLiveWallpaper() {
+export function createMacLiveWallpaper() {
   if (wallWindow.length > 0) {
     return
   }
@@ -15,19 +15,19 @@ export function createLiveWallpaper() {
   displays.forEach(async (display, index) => {
     wallWindow.push(
       new BrowserWindow({
-        show: false,
-        type: 'desktop',
-        focusable: false, //窗口是否可以获取焦点
+        show: false, // 是否显示窗口
+        type: 'desktop', // 设置窗口类型为桌面窗口
+        focusable: false, // 窗口是否可以获取焦点
+        frame: false, // 是否显示边缘框
         x: display.bounds.x,
         y: display.bounds.y,
         width: display.bounds.width,
         height: display.bounds.height,
-        frame: false, //是否显示边缘框
-        // fullscreen: true, //是否全屏显示
+        // fullscreen: true, // TODO 是否全屏显示, 会导致窗口无法显示
         webPreferences: {
-          nodeIntegration: true, //赋予此窗口页面中的JavaScript访问Node.js环境的能力
-          webSecurity: true, //可以使用本地资源
-          contextIsolation: false, //是否使用上下文隔离
+          nodeIntegration: true, // 赋予此窗口页面中的JavaScript访问Node.js环境的能力
+          webSecurity: true, // 可以使用本地资源
+          contextIsolation: false, // 是否使用上下文隔离
         },
       }),
     )
